Fixed type hint for total_cost_calculation

+++ app/calculator.py
@@ -226,7 +226,7 @@ class Calculator():
         return state
 
     def total_cost_calculation(self, start_date: date, start_time: time, end_time: datetime,
-                               start_state: int, base_price: int, power: int, capacity: int,
+                               start_state: int, base_price: float, power: float, capacity: float,
                                postcode: str, solar_energy: float = 0) -> float:
         state = self.get_state(postcode)
         total_holiday_peak = 0
